+++ src/GafferImageUI/ImageGadget.cpp
@@ -258,7 +258,7 @@ void ImageGadget::setSoloChannel( int index )
 		// channels.
 		dirty( TilesDirty );
 	}
-	requestRender();
+	Gadget::dirty( DirtyType::Render );
 }
 
 int ImageGadget::getSoloChannel() const
@@ -276,7 +276,7 @@ void ImageGadget::setClipping( bool clipping )
 	}
 	else
 	{
-		requestRender();
+		Gadget::dirty( DirtyType::Render );
 	}
 }
 
@@ -294,7 +294,7 @@ void ImageGadget::setExposure( float exposure )
 	}
 	else
 	{
-		requestRender();
+		Gadget::dirty( DirtyType::Render );
 	}
 }
 
@@ -313,7 +313,7 @@ void ImageGadget::setGamma( float gamma )
 	}
 	else
 	{
-		requestRender();
+		Gadget::dirty( DirtyType::Render );
 	}
 }
 
@@ -343,7 +343,7 @@ void ImageGadget::setDisplayTransform( ImageProcessorPtr displayTransform )
 
 	dirty( TilesDirty );
 	m_shaderDirty = true;
-	requestRender();
+	Gadget::dirty( DirtyType::Render );
 }
 
 ConstImageProcessorPtr ImageGadget::getDisplayTransform() const
@@ -369,7 +369,7 @@ void ImageGadget::setLabelsVisible( bool visible )
 		return;
 	}
 	m_labelsVisible = visible;
-	requestRender();
+	Gadget::dirty( DirtyType::Render );
 }
 
 bool ImageGadget::getLabelsVisible() const
@@ -391,7 +391,7 @@ void ImageGadget::setPaused( bool paused )
 	}
 	else if( m_dirtyFlags )
 	{
-		requestRender();
+		Gadget::dirty( DirtyType::Render );
 	}
 }
 
@@ -490,7 +490,9 @@ void ImageGadget::dirty( unsigned flags )
 	}
 
 	m_dirtyFlags |= flags;
-	requestRender();
+	Gadget::dirty(
+		( FormatDirty | DataWindowDirty ) ? DirtyType::Bound : DirtyType::Render
+	);
 }
 
 const GafferImage::Format &ImageGadget::format() const
@@ -763,7 +765,7 @@ void ImageGadget::updateTiles()
 			ParallelAlgo::callOnUIThread(
 				[thisRef] {
 					thisRef->m_renderRequestPending = false;
-					thisRef->requestRender();
+					thisRef->Gadget::dirty( DirtyType::Render );
 				}
 			);
 		}