+++ src/GafferUI/Gadget.cpp
@@ -56,8 +56,56 @@ using namespace std;
 
 GAFFER_GRAPHCOMPONENT_DEFINE_TYPE( Gadget );
 
+//////////////////////////////////////////////////////////////////////////
+// Gadget::Signals
+//
+// We allocate these lazily because they have a significant overhead
+// in both memory and construction time, and for many Gadget
+// instances they are never actually used.
+//////////////////////////////////////////////////////////////////////////
+
+struct Gadget::Signals : boost::noncopyable
+{
+
+	VisibilityChangedSignal visibilityChangedSignal;
+	RenderRequestSignal renderRequestSignal;
+
+	ButtonSignal buttonPressSignal;
+	ButtonSignal buttonReleaseSignal;
+	ButtonSignal buttonDoubleClickSignal;
+	ButtonSignal wheelSignal;
+
+	EnterLeaveSignal enterSignal;
+	EnterLeaveSignal leaveSignal;
+	ButtonSignal mouseMoveSignal;
+
+	DragBeginSignal dragBeginSignal;
+	DragDropSignal dragEnterSignal;
+	DragDropSignal dragMoveSignal;
+	DragDropSignal dragLeaveSignal;
+	DragDropSignal dragEndSignal;
+	DragDropSignal dropSignal;
+
+	KeySignal keyPressSignal;
+	KeySignal keyReleaseSignal;
+
+	// Utility to emit a signal if it has been created, but do nothing
+	// if it hasn't.
+	template<typename SignalMemberPointer, typename... Args>
+	static void emitLazily( Signals *signals, SignalMemberPointer signalMemberPointer, Args&&... args )
+	{
+		if( !signals )
+		{
+			return;
+		}
+		auto &signal = signals->*signalMemberPointer;
+		signal( std::forward<Args>( args )... );
+	}
+
+};
+
 Gadget::Gadget( const std::string &name )
-	:	GraphComponent( name ), m_style( nullptr ), m_visible( true ), m_enabled( true ), m_highlighted( false ), m_toolTip( "" )
+	:	GraphComponent( name ), m_style( nullptr ), m_visible( true ), m_enabled( true ), m_highlighted( false ), m_layoutDirty( false ), m_toolTip( "" )
 {
 	std::string n = "__Gaffer::Gadget::" + boost::lexical_cast<std::string>( (size_t)this );
 	m_glName = IECoreGL::NameStateComponent::glNameFromName( n, true );
@@ -102,7 +150,9 @@ void Gadget::setStyle( ConstStylePtr style )
 		{
 			const_cast<Style *>( m_style.get() )->changedSignal().connect( boost::bind( &Gadget::styleChanged, this ) );
 		}
- 		requestRender();
+		// Style affects the bounding box of text,
+		// so we need Layout rather than just Render.
+		dirty( DirtyType::Layout );
 	}
 }
 
@@ -137,9 +187,13 @@ void Gadget::setVisible( bool visible )
 	if( !p || p->visible() )
 	{
 		emitDescendantVisibilityChanged();
-		visibilityChangedSignal()( this );
+		Signals::emitLazily( m_signals.get(), &Signals::visibilityChangedSignal, this );
 	}
- 	requestRender();
+	Signals::emitLazily( m_signals.get(), &Signals::renderRequestSignal, this );
+	if( p )
+	{
+		p->dirty( DirtyType::Layout );
+	}
 }
 
 void Gadget::emitDescendantVisibilityChanged()
@@ -153,7 +207,7 @@ void Gadget::emitDescendantVisibilityChanged()
 			continue;
 		}
 		(*it)->emitDescendantVisibilityChanged();
-		(*it)->visibilityChangedSignal()( it->get() );
+		Signals::emitLazily( (*it)->m_signals.get(), &Signals::visibilityChangedSignal, it->get() );
 	}
 }
 
@@ -178,7 +232,7 @@ bool Gadget::visible( Gadget *relativeTo ) const
 
 Gadget::VisibilityChangedSignal &Gadget::visibilityChangedSignal()
 {
-	return m_visibilityChangedSignal;
+	return signals()->visibilityChangedSignal;
 }
 
 void Gadget::setEnabled( bool enabled )
@@ -188,7 +242,7 @@ void Gadget::setEnabled( bool enabled )
 		return;
 	}
 	m_enabled = enabled;
-	requestRender();
+	dirty( DirtyType::Render );
 }
 
 bool Gadget::getEnabled() const
@@ -218,7 +272,7 @@ void Gadget::setHighlighted( bool highlighted )
 	}
 
 	m_highlighted = highlighted;
- 	requestRender();
+	dirty( DirtyType::Render );
 }
 
 bool Gadget::getHighlighted() const
@@ -231,7 +285,10 @@ void Gadget::setTransform( const Imath::M44f &matrix )
 	if( matrix!=m_transform )
 	{
 		m_transform = matrix;
- 		requestRender();
+		if( auto p = parent<Gadget>() )
+		{
+			p->dirty( DirtyType::Layout );
+		}
 	}
 }
 
@@ -255,6 +312,7 @@ Imath::M44f Gadget::fullTransform( const Gadget *ancestor ) const
 
 void Gadget::render() const
 {
+	bound(); // Updates layout if necessary
 	for( int layer = (int)Layer::Back; layer <= (int)Layer::Front; ++layer )
 	{
 		renderLayer( (Layer)layer, /* currentStyle = */ nullptr );
@@ -311,14 +369,35 @@ void Gadget::renderLayer( Layer layer, const Style *currentStyle ) const
 	}
 }
 
-void Gadget::requestRender()
+void Gadget::dirty( DirtyType dirtyType )
 {
 	Gadget *g = this;
 	while( g )
 	{
-		g->renderRequestSignal()( g );
+		if( dirtyType == DirtyType::Layout )
+		{
+			g->m_layoutDirty = true;
+		}
+		Signals::emitLazily( g->m_signals.get(), &Signals::renderRequestSignal, g );
+		if( dirtyType == DirtyType::Bound )
+		{
+			// Bounds changes in children require layout updates in parents.
+			dirtyType = DirtyType::Layout;
+		}
 		g = g->parent<Gadget>();
 	}
+}
+
+void Gadget::updateLayout() const
+{
+}
+
+void Gadget::requestRender()
+{
+	// `requestRender()` has been deprecated and replaced by `dirty()` because
+	// it didn't provide the fine-grained control we need. Where extension code
+	// is still using it, we must assume the worst and dirty the layout.
+	dirty( DirtyType::Layout );
 }
 
 void Gadget::doRenderLayer( Layer layer, const Style *style ) const
@@ -332,7 +411,13 @@ bool Gadget::hasLayer( Layer layer ) const
 
 Imath::Box3f Gadget::bound() const
 {
-	Box3f result;
+	if( !m_layoutDirty )
+	{
+		return m_bound;
+	}
+
+	updateLayout();
+	m_bound = Box3f();
 	for( ChildContainer::const_iterator it=children().begin(); it!=children().end(); it++ )
 	{
 		// cast is safe because of the guarantees acceptsChild() gives us
@@ -343,9 +428,11 @@ Imath::Box3f Gadget::bound() const
 		}
 		Imath::Box3f b = c->bound();
 		b = Imath::transform( b, c->getTransform() );
-		result.extendBy( b );
+		m_bound.extendBy( b );
 	}
-	return result;
+
+	m_layoutDirty = false;
+	return m_bound;
 }
 
 Imath::Box3f Gadget::transformedBound() const
@@ -362,7 +449,7 @@ Imath::Box3f Gadget::transformedBound( const Gadget *ancestor ) const
 
 Gadget::RenderRequestSignal &Gadget::renderRequestSignal()
 {
-	return m_renderRequestSignal;
+	return signals()->renderRequestSignal;
 }
 
 std::string Gadget::getToolTip( const IECore::LineSegment3f &position ) const
@@ -377,77 +464,86 @@ void Gadget::setToolTip( const std::string &toolTip )
 
 Gadget::ButtonSignal &Gadget::buttonPressSignal()
 {
-	return m_buttonPressSignal;
+	return signals()->buttonPressSignal;
 }
 
 Gadget::ButtonSignal &Gadget::buttonReleaseSignal()
 {
-	return m_buttonReleaseSignal;
+	return signals()->buttonReleaseSignal;
 }
 
 Gadget::ButtonSignal &Gadget::buttonDoubleClickSignal()
 {
-	return m_buttonDoubleClickSignal;
+	return signals()->buttonDoubleClickSignal;
 }
 
 Gadget::ButtonSignal &Gadget::wheelSignal()
 {
-	return m_wheelSignal;
+	return signals()->wheelSignal;
 }
 
 Gadget::EnterLeaveSignal &Gadget::enterSignal()
 {
-	return m_enterSignal;
+	return signals()->enterSignal;
 }
 
 Gadget::EnterLeaveSignal &Gadget::leaveSignal()
 {
-	return m_leaveSignal;
+	return signals()->leaveSignal;
 }
 
 Gadget::ButtonSignal &Gadget::mouseMoveSignal()
 {
-	return m_mouseMoveSignal;
+	return signals()->mouseMoveSignal;
 }
 
 Gadget::DragBeginSignal &Gadget::dragBeginSignal()
 {
-	return m_dragBeginSignal;
+	return signals()->dragBeginSignal;
 }
 
 Gadget::DragDropSignal &Gadget::dragMoveSignal()
 {
-	return m_dragMoveSignal;
+	return signals()->dragMoveSignal;
 }
 
 Gadget::DragDropSignal &Gadget::dragEnterSignal()
 {
-	return m_dragEnterSignal;
+	return signals()->dragEnterSignal;
 }
 
 Gadget::DragDropSignal &Gadget::dragLeaveSignal()
 {
-	return m_dragLeaveSignal;
+	return signals()->dragLeaveSignal;
 }
 
 Gadget::DragDropSignal &Gadget::dropSignal()
 {
-	return m_dropSignal;
+	return signals()->dropSignal;
 }
 
 Gadget::DragDropSignal &Gadget::dragEndSignal()
 {
-	return m_dragEndSignal;
+	return signals()->dragEndSignal;
 }
 
 Gadget::KeySignal &Gadget::keyPressSignal()
 {
-	return m_keyPressSignal;
+	return signals()->keyPressSignal;
 }
 
 Gadget::KeySignal &Gadget::keyReleaseSignal()
 {
-	return m_keyReleaseSignal;
+	return signals()->keyReleaseSignal;
+}
+
+Gadget::Signals *Gadget::signals()
+{
+	if( !m_signals )
+	{
+		m_signals.reset( new Signals );
+	}
+	return m_signals.get();
 }
 
 Gadget::IdleSignal &Gadget::idleSignal()
@@ -470,7 +566,7 @@ Gadget::IdleSignal &Gadget::idleSignalAccessedSignal()
 
 void Gadget::styleChanged()
 {
-	requestRender();
+	dirty( DirtyType::Layout );
 }
 
 void Gadget::parentChanged( GraphComponent *oldParent )
@@ -479,10 +575,10 @@ void Gadget::parentChanged( GraphComponent *oldParent )
 
 	if( Gadget *oldParentGadget = IECore::runTimeCast<Gadget>( oldParent ) )
 	{
-		oldParentGadget->requestRender();
+		oldParentGadget->dirty( DirtyType::Layout );
 	}
 	if( Gadget *parentGadget = parent<Gadget>() )
 	{
-		parentGadget->requestRender();
+		parentGadget->dirty( DirtyType::Layout );
 	}
 }