+++ src/GafferUI/GraphGadget.cpp
@@ -1257,7 +1257,7 @@ bool GraphGadget::dragMove( GadgetPtr gadget, const DragDropEvent &event )
 		offsetNodes( m_scriptNode->selection(), pos - m_lastDragPosition );
 		m_lastDragPosition = pos;
 		updateDragReconnectCandidate( event );
- 		requestRender();
+		dirty( DirtyType::Render );
 		return true;
 	}
 	else
@@ -1265,7 +1265,7 @@ bool GraphGadget::dragMove( GadgetPtr gadget, const DragDropEvent &event )
 		// we're drag selecting
 		m_lastDragPosition = V2f( i.x, i.y );
 		updateDragSelection( false, event.modifiers );
- 		requestRender();
+		dirty( DirtyType::Render );
 		return true;
 	}
 
@@ -1399,12 +1399,12 @@ bool GraphGadget::dragEnd( GadgetPtr gadget, const DragDropEvent &event )
 		}
 		m_dragReconnectCandidate = nullptr;
 		m_dragMergeGroupId++;
- 		requestRender();
+		dirty( DirtyType::Render );
 	}
 	else if( dragMode == Selecting )
 	{
 		updateDragSelection( true, event.modifiers );
- 		requestRender();
+		dirty( DirtyType::Render );
 	}
 
 	return true;
@@ -1644,6 +1644,21 @@ NodeGadget *GraphGadget::addNodeGadget( Gaffer::Node *node )
 	{
 		return nullptr;
 	}
+
+	// The call to `addChild( nodeGadget )` will uniquefy the name
+	// with respect to all our other NodeGadgets. This can have
+	// significant overhead when graphs get large, which we can
+	// avoid by making sure the name is unique already. The node
+	// name fits the bill, and is already conveniently available
+	// in `InternedString` form. Note that we are deliberately not
+	// synchronising the names in the case of them changing in the
+	// future, and provide no guarantees about NodeGadget names in
+	// general.
+	// \todo It may be better to modify GraphComponent to allow nameless
+	// children and then use them here. This would let us avoid _all_
+	// uniquefication overhead and would also be useful for ArrayPlug
+	// and Spreadsheet::RowsPlug children.
+	nodeGadget->setName( node->getName() );
 
 	addChild( nodeGadget );